Allow cpusets to use all cpus but disabling the cpu hotplug of sprdemand
Setting cpuset.cpus to 0-3 would only work if all 4 cores are online
which is not always the case.

The sitch would be somewhat the cpuset.cpus would always be set to 0
meaning only one core will be used even when all core are re-onlined again.

This is a band-aid patch, and we need a better solution since hotplugging
may actually help in powersaving.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>