Remove overlaps at common.mk, FMRadio and dither; and add some stuffs
I'll add the new browser Gello, and the Snap camera app which is nice since
it supports panorama! We've got Live Wallpapers too and two more properties
from the lollipop (and from stock) build.

The FMRadio still doesn't work, and the overlaps -- they are present in
sprd-common/common.mk and in this -- are redudant so remove them.

The libdither is included when USE_SPRD_DITHER is defined so it is
removed.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>