Modify permissions for the userspace apps/binaries
OMX de/encoder initialization should soon fail accessing the
/dev/sprd_vsp character device. This is resolved by this patch
yet I don't know why.

Later on on camera init, it would also complain about
not being able to write on
/sys/devices/virtual/camera/rear/Cam_antibanding. Also looking
in the init files, it seems to set the appropiate groups
but with the wrong command. Changing 'chmod' to 'chown' should fix it.

There's also permission problems regarding pin_switch from audio init.
I think this is because the tinyalsautils was pushed to the audio libs
instead of the system code (by using the patches from cm-12.1)
so the acting group of the function accessing these is
now in audio group not the system group.

NOTE:
- The ueventd files are based on j3xnlte device tree.
- Marking /dev/sprd_gsp with 777 (global permissions) could become a possible security hole.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>