Disable healthd during charging; Don't set powersave governor during charing
healthd when activated on charging while the phone is turned off,
shows its own charging animation apart from the
lpm binary from stock firmware.
Also it seems that healthd alone cannot write to the framebuffer.
In other terms, healthd only draw its own when and ony when lpm is running.

---
The powersave governor disables the cpu hotplugging that comes along with sprdemand.
The other 3 cores are not needed in the lpm(low power mode)
and sprdemand governor also scales down to the minimum like powersave

In short, setting to powersave is not that beneficial.

Although most of the time during charging,
the phone enters a deep sleep state, causing most of the cpus to shutdown.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>