Switch over to media_codecs_google_video_le.xml from media_codecs_google_video.xml
As said on commit 17e95f40a534101e2df48f077a8029430adef6c3 upstream

	This can be used by low-end devices

Why not try this instead of the no "_le" xml?

It might benefit the mid-ness of the device.
SM-G355H is not that of a high-end nor is that low to be called as
low-end but somewhere in between.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>