Go back to media_profiles.xml from stock firmware
As I thought, there's something funky on the video recording.

First, there's an option to record on 720p setting which is
       impossible on this hardware.

Second, the audio recorded is played back at a faster rate
       causing it to rise in pitch.

Third, the video has some serious A/V desynchronization,
       the audio has already ended while the video is still playing.

I don't know how the xml config is written to just simply fix the
problem so I brought back the file from stock.

But I suspect the problem to be on the audio encoding bitrate or on
some wrong n bit used for encoding (i.e. s32le instead of s16le).

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>