Use the phone's partition sizes; Add back the Camera config from cm 12.1
To prevent build errors about the recovery image is larger than the
board's capacity, let's make the size correspond to the values on the device.

By using the start sector and end sector, and the sector size stat of sgdisk
You can compute the size in bytes by the formula:
( End_sector - Start_sector + 1 ) * sector_size_in_btyes

----
I faced some compilation error in the linking stage for camera2.sc8830
because of the missing symbols from libisp.

Mysteriously after copy-paste'ing the config from cm 12.1, it works.
I've removed the false-valued flags that seems to be the default option.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>