Rearrange rootdir for initrd
I saw that the init.rc is actually provided by us but it does not
differ that much to the one given in the source. The only notable difference is the
ril-daemon services.
I re-arranged it so the init.rc can be as close to the original and moved everything
else to other init files.

There's also the init.storage.rc that I think is not used anymore; it's just there
without any file importing/referencing to it.

There's that 'setenv' scattered in init.kanas3g_base.rc. I think it would be better
if it is done once at the start of the init with the 'export' command.

init.swap.rc is now imported and read at init.kanas3g.rc

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>