Try reading current_now instead of current_max; Convert mA to uA
The charger driver always reports at milliamperes but Android
sort of expects this to be in microamperes.

Android also reads the current_max file for the current.

The problem is that the charger driver's current_max seems to be
an arbritrarily defined constant, pegged at 2A or 2000mA. The
one that has been changing is the current_now.

It seems that Android needs to read that file and when it fails
go back to accessing current_max.

To fix the mA and uA differrences, the value just simply needed to be
multiplied by 1000 before being sent through IPC.

This adds a patch and modifies the resource overlay for Keyguard.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>