Disable cpusets, go back to the "old sched policy"
The intention of ActivityManager using cpusets is to
distribute the tasks across the big.LITTLE
processor cores, which the device doesn't have.

But the processor has hotplugs and cores
which doesn't play nicely with cpusets.
This causes only one core to be active
with absurdly high load. It ticks off the
hotplug logic too by making it think it
needs to online more cores.

Thus, leading up to a scenario where
CPU0's (first core) utilization is 100%,
all other 3 cores are online but not sharing the
load with CPU0, and the performance crawls.

This reverts the commit I added earlier which is named:
"Allow cpusets to use all cpus but disabling the cpu hotplug of sprdemand"

Yep, that "but" should actually be "by". My bad.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>