Remove the SPRD_HARDWARE macro in "needs-flush-before-disable quirk" patch
It seems to be generic enough and not likely to hurt anyone having this guard
off.

Plus I forgot to define this macro on BoardConfig.mk making that patch useless
in the first place. This commit makes it work.

That patch also seems to smoothen out playback...

I tested this on youtube.
Before the patch, there seems to be breaks/glitches on the playback.
It's sound something like playing a scratched cd.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>