Switch to opensource modemd
This is to remedy (not prevent or fix) the modem failure that occurs when having
two sims in dsds (dual sim) mode. It allows modemd to peacefully reset the modem when it fails.

This adds the necessary system properties and init services for this modemd version to work.
I've also added commands to support other refnotify binaries.

I am sorry if this commit *BREAKS* the previous version of modemd from android_vendor_samsung_kanas.

This commit is split to two. Check the other on android_vendor_samsung_kanas.

This also requires that source code to be in android_hardware_sprd.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>