Depend on the scx35-common repo; simplify init files
This would help trim a lot of code and maintenance.

Our device is fortunate that **most** (but not all) of the files and code needed
are already in that repository and actually does work with this device.

What was done to simply remove the same pieces of code.
Using the core33g repo as a reference, I've altered and added some init files like for rild.

Even though it does work, it needs a bit of adjustments like the extra
codecs for aac, the always clicking audio, bluetooth not pairing.

Expect the next few patches to fix and tweak these nuisances and some pollishing.

Note that the patch included is needed to make these work.
I'm not sure whether it should be PR'd since kanas3g/kanas isn't exactly scx35.
I don't really know the specifications since I'm just a simple user of this device.

Change-Id: I9c4d23dcf008f17ff74c9596a19789a6ea7b7eac
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>