Force xdpi and ydpi values to be at 240.0
With recent changes in Android Nougat there seems to be
a problem with the xdpi and ydpi values. It's actually right.

It's too right that it causes problems with applications
and, worse, system services like the HIC of the Keyguard service
that maintains the lock screen.

In Lollipop and Marshmallow, it's usually according to the screen dpi so set it to be 240 for now.

Note: THIS IS A WORKAROUND NOT A FIX and requires a patch to hwcomposer for this to work.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>