kanas: Make gralloc use 3 buffers to enable page flipping
Fixes the xdpi/ydpi issue and makes it set the real values.

I apologize if you didn't caught on the sarcasm on the previous commit:
"Force xdpi and ydpi values to be at 240.0". 25.4 is not a real/pratical dpi value.
That's too small.

This, in effect, also reverts the aforementioned patch.

Change-Id: I044a9396fff532460315756b5aafd6dadcc4dc12
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>