kanas: Enable Dexpreopt
The issue with dexpreopt is because the PRODUCT_BOOT_JARS is out of order.
This causes invalid memory access on runtime triggered by whatever is
the first element in PRODUCT_BOOT_JARS. In my case, it's when the wifi card
successfully connects to an AP.

Shifting the include directives is a simple solution to avoid that.

Change-Id: If3865ae0ba36fd3f87292c979f52df9d3040e64c
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>