kanas: libsensors: Use Accelerometer's SMD implementation
It turns out that the BOSCH BMA254/BMA255 have a hardware support for
Significant Motion Detection as said in its datasheet and bma2x2 kernel driver.

Though the Samsung driver does not expose this, we can work around this on the kernel
side then expect input events on the userspace side.

Besides, the BMA254 accelerometer can wake the device from deep sleep and it's using about
0.2 mA which is way lower than the 1 mA stated in Android sensor documentation.

Change-Id: I2b24765b8175babf54f36f84b1878a5faeeab68c
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>