kanas: libsensors: Backport user-made Android sensors HAL module
This interfaces to the kernel driver and communicates through HAL 1.3
instead of HAL1.0.

It was originally intended to avoid using the sensors wrapper library
but I figured out that it can be extended to support SMD on the hardware level.

What I mean about "on the hardware level" bit is that the SMD is done on the
accelerometer instead of being emulated on the software level. This would
need a kernel patch though.

With this we can use Doze in Marshmallow, Deep Doze in Nougat or
whatever's in Android O and P.

And oh, this is the marshmallow commit.

Change-Id: Ie6e08cf9c522e1166711f4c92c2bc997aa3efaf4
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>