kanas: Adapt their patching mechanism
I really liked how their mechanism works in patching and
it also addresses the issue of having a set of patches.

I also included my version with requires bash which is a simplified
version. Unfortunately, that is not actually equivalent to the original
version that can run on sh. One example is that it won't recognize
any *.patch files on the patches/ directory since the bash version assumes
that the root of the source tree can not be under a git repository.

This commit also adapts the current patches to this relatively new mechanism.
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>

Change-Id: I9afa6461f1a0a306203f9351997d992770ab151f