kanas: Allow page flipping in gralloc
Well, not really noticable in Android Marshmallow and older, the xdpi and ydpi
are set to 25.4. It's too small and would cause some apps to go bonkers
in scaling visual elements like images.

Take a look at this:
https://github.com/inorichi/tachiyomi/issues/1539

This sets the flag that tells gralloc that the device uses a larger framebuffer
which allows the FBIOPUT_VSCREENINFO ioctl to succeed.

When it fails, the ioctl call seems to set the width and height to be like:
width = xres and height = yres.

This solves the issue with x/ydpi values and allows page flipping
in the gralloc module. :)

Change-Id: I9740139973741297030a58e0f4ad16efc80ad10d
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>