kanas: powerhal: Limit GPU frquency
I figured that limiting GPU frequency, if possible, could bring down
power consumption further. I was able to get as low as 8%/hr consumption
while reading comics under the powersave mode (which limits GPU to 64MHz).

I'm still experimenting with this and
the fact that it needs the mali platform code to support scaling
to frequencies as low as 64MHz and to expose 4 cores(jobs?) is a let down.

Change-Id: I05abf58bfde43d92ad4bdda33ac9c84e446e081a
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>