kanas: Automatically apply patches when loading vendorsetup.sh
This usually activates when `source build/envsetup.sh` is run.
Honestly, it bothers me so much that most of my problems (i.e. compile errors)
is due to the patches that I forgot to patch.

I take the patches for granted that I totally forgot that a single `repo sync`
can revert the trees of the patches.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>