Revert "kanas: Use the pinner service"
Okay so, We want our 69 MB back! Though those ain't much for 700 MB of RAM,
it is when you start thinking it relative to the free space and the usage of the
system_server process.

After testing diffrent configurations, the memory usage did not drop.
It can be observed that, relative to the version without the pinnerservice config,
the system_server process has increased almost 100% of it's memory usage.

The extra memory used, albeit can be shared, was not usually shared if not at all.
This is also seems to help to trigger the LMK quite frequently.

In conclusion, this brings almost no benefit.

This reverts commit 93a0c90c6a580cebd674db953070a3a3ff6bbf02.