kanas: overlays: Reduce LMK values
Unfortunately, this device does seem to sit on the edge at what the Activity Manager thinks to be a high memory device (~700MB).
In fact, at least in the yac kernel, the total memory is ~707MB.
This causes AM to compute a scale of 1 from any real number in 0 to 1 inclusive.
In effect, the LMK values assigned are the highest the developers had set.

It has a funny effect where the last ~200 MB of the RAM is unused since LMK works early as the free memory goes below ~180 MB.

Though the algorithm can be changed, opt to use the config_lowMemoryKillerMinFreeKbytesAbsolute setting they created.

Change-Id: I63272cf2aa80e9c73f3b7fa42a6c02300e55e7b7
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>