kanas: gps: A wrapper to bridge API differences with the GPS HAL
This device GPS uses one from Marvell with some customizations from Spreadtrum.
Though the HAL works for Android 6, due to recent changes in Android 7
the HAL won't work anymore.

Thankfully, of Google, the old GPS interface and callback structs are still possible
to be used but it needs additional steps to do so. Thus, this wrapper was created.

To avoid plagarisms, this was made from:
https://github.com/LineageOS/android_device_motorola_omap4-common/commit/f7a128e12aa79cc018c44ff929bd7d44eb95d7eb
with some changes with it's logging (I made sure to poke fun of how Marvell's
(or maybe Spreadtrum's) libgps does it's logging).

Those additional steps are the following:
	1. Use the older GPSCallback_v1 when calling init() of the HAL
	2. Use the older AGpsInterface_v1 after returning the agps interface.

The first (1) is to verify the struct size check in the GPS HAL, with the newer GPSCallback
having a larger size than that of the older struct, which why HAL initialization fails.
What I mean about HAL initialization, it DOES NOT INCLUDE GPS chip init!
That is a separate issue solved on some other commit.

The second (2) should fix the agps that was once broken on Android 6: (from the vendor side)
9bfb13ddf8e6049c2bddb6736d3f8ba78cb44859 gps: Disable agps
That issue lies around the GnssLocationProvider stuff that mistakes the version of the
AGpsInterface struct the GPS HAL returns. It thinks that the AGpsInterface struct returned
is the newer one! The problem lies where it tries to access an element (a function) in the
new struct that is non-existent on the older struct. It should have been an illegal access.
But like the undefinability in C, we go to some random part of the code and fail there.
Now, the wrapper makes sure the GnssLocationProvider JNI properly determines the right struct
and calls the appropriate function.

Change-Id: I8d5fdbec9812ee75e875376f2af21b93b17d2a42
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>