kanas: gps: fix derp merging
This is to fix the wrong commit at:
kanas: gps: A wrapper to bridge API differences with the GPS HAL

There's a bunch of untested last minute changes like removing the loglevel
variable and using the gps.default.so name. This commit should fix them all at once.

Speaking of which the last tested version had a make-shift lexer and parser
for the mrvl_gps_platform.conf file which is used to set the loglevel from the
file.

Change-Id: I73b5079fbb6e039879371345690de475508c5b07
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>