kanas: init: automatically start refnotify
...Why is this not on?

We may go to problems where the modem queue become so full that it triggers an assert
which needs to do a modem reset.

Though about that, it's just my speculation and I have no intention of testing that
since it may take about a month for it to activate.

Change-Id: I6b446b4d1d2f1600767a560b0ddaae7b4fbedc26
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>