kanas: ril: Block all calls to getModemActivityInfo and getCellInfoList
getModemActivityInfo is something added in Android 6 and this device's ril does not
support it. One thing I noticed is that the Android system would regularly poll
the modem with this then would hold a wakelock for about 2 seconds just to
receive an error from the rill. This happends frequently enough that RILJ wakelock
will eventually show up as one of the "most active" wakelock. To try and minimize
that, we fail fast and reject the message, not passing it down the RIL.

Regarding the getCellInfoList, this is called less frequently. Though it is supported
by the RIL, or at least recognized. Somehow, all calls returns an E_GENERIC_FAILURE error.
I don't know why it's called that often so I disabled it for now.

Change-Id: Ic5bb62d28e46e54630bd1a4ce0cc078de754ff42
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>