kanas: set NUM_FRAMEBUFFER_SURFACE_BUFFERS to 2 instead of 3
This flag is set to 3 as it prevents screen tearing but it seems that
it just need to be set to any number. Without any knowledge of which to set,
I set it back to the "default" number (2) as defined in
frameworks/native/services/surfaceflinger/DisplayHardware/FramebufferSurface.cpp
right about line 45.

So far nothing seems to change on the UX side after a clean build, no tearing
of some sort.

Change-Id: I578e33d9f7d7174df90676ad90b4e87922ef724c
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>