kanas: system.prop: Remove config.disable_atlas flag
This is actually a relic from the BlissPop Build I did few years back
when I was exploring the possiblity of building my own ROM. It aims to
conserve memory by disabling the AssetAtlas (?) service.

That service allows preloading graphical assets onto memory and provides
a mechanism for other services and apps to share the same assets reducing
multiple copies on memory. This has the side effect of increasing performance
but the service itself and the preloaded assets may be detrimental to the
768... err 707 MB of RAM the device has.

Well, it doesn't seem to impact it that much plus it does not even work
at all. config.disable_atlas is a flag that does not exist in AOSP but it
does on some derivatives like LOS(or CM). And, the way it is coded.
It needs the flag config.disable_noncore to be set to true.
Meanwhile, setting config.disable_noncore to true will break the system.

The line in question goes like this in
frameworks/base/services/java/com/android/server/SystemServer.java:
`if (!disableNonCoreServices && ZygoteInit.PRELOAD_RESOURCES && !disableAtlas) {`

In short, I'm removing this since it does not do anything.

Change-Id: I7e7ea13da2dbbde73cefe1653666c95e58d5cd3b
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>