kanas: enable PIP mode in nougat
Enabling the PIP mode in nougat is a bit of a joke since it has a quite wanky
requirement of sending keycode 171 for it to activate. Once active, it sits
on the top left corner and it cannot be brought back to full size.

Revert until a solution, other than going to Android O, is found.

Change-Id: I82285e9a7883e1b1c7ccafdc8153b0e69aef2ad2
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>