kanas: ril: filter Cell Broadcast message IDs to be set.
The modem seem to only support 20 different ids. The observed problem here is that on Android 7 and above the ids to be set can easily exceed 20. And, that's seem to be the case on default settings.

I've also noticed this to have a miniscule impact on the power usage, specially on 3G.

The whole RIL also does not seem to unset the ids when disabled from Android even though the list shows that the id has been removed. This means once a id is enabled, it can't be disabled from the RIL (or through Android). One way to unset is by issuing an AT command from elsewhere.

This approach limits the set ids by matching the id range to a predefined list of ids. I've copied the default set of ids from Android 5.1 with 5 slots to spare. The list must be sorted due to the implementation of the filter.

4352,4353,4354,4356,4370,4371,4372,4373,4374,4375,4376,4377,4378,4379,4383

Change-Id: I5239bc5e81c005c98eb553de338b8190f05b4fc9
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>