kanas: Add "/dev/ion (deleted)" as one of the whitelisted paths in zygote; Do not use scx35-common's custom gps.h header
The first part basically screams about this devoloper's incompetence to
understand why does the fd to /dev/ion is being "deleted" causing
zygote to fail due to security purposes. Of course, the system
does not boot.

The second part pertains to the gps.h being used as a part of the common config
from scx35-common. We're using the gps hal wrapper so it shouldn't matter if
we are using the custom or the original header.

Change-Id: Ib32533017abce532cf3ce0361c6d8fe2e8fca0e8
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>