kanas: patches: Mitigate HIDL incompatibility with the gralloc module
After some research, it can be observed that the HIDL issue regarding using
sprd gralloc module in passthrough mode can be, in fact, mitigated using
some cheap hack, rather than the rewriting the private_module_t/private_handle_t.
Honestly, the latter may not even work.

Overall there are two problems: FB memory is free'd immediately after being
allocated, and the gralloc module unconventionally uses a -1 fildes which
ticks of the security mechhanism of binder.

The hacks are pretty straight forward.

First, since FB memory is almost
always (99%) being used you may pretty much assume that no one will
call free on it, so we never free it at all.

Second, the EGL module and gralloc module work in tandem.
The -1 fildes seems to be used as an indicator by the EGL module
to determine that the buffer is from an FB so simply creating
a dummy FD will not work. But, it can be safely mitigated by
imediately closing and invalidating the fd after
it passes the binder check. This usually
occurs on the register buffer function.

With these hacks we won't need the hack injecting the Gralloc1/0 code in libui.

Change-Id: I91dc11d8c8814c23328ad19ebaafd22045a3e915
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>