kanas: patches: Enable USE_OVERLAY_COMPOSER_GPU
I think, I have fixed this stuff. Overlay composer is the one responsible
of ofloading the blending to HW (if my understanding is correct)
and ,based on my unprecise and inaccurate approximation, gives a
1-2 fps boost in the general case.

This was broken in the first place. First because of the patch that adds back
Gralloc1/0 support in libui destroys the symlink that leads to the right
includes folder.
Second, the ANativeWindow is split to a libnativewindow library which
is needed.
Third is that this still uses the older GraphicBuffer constructor that
doesn't exist in the exact signature as before.

Change-Id: I05824e51b70e4a1bf3abc9a414d205bd0eb9d000
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>