kanas: patches: Create a TARGET_USE_3_FRAMEBUFFER flag
For some reason, BOARD_GLOBAL_CFLAGS doesn't work anymore.
We need to define USE_3_FRAMEBUFFER on gralloc so that it sets
the right xdpi annd ydpi.

If this is not set properly, certain apps (tachiyomi) which depends on
the xdpi and ydpi values on scaling may display images relatively smaller
the an average adult household cockroach.

Change-Id: Ic742b42d640cc67dd3dc297d7c5ee1b1a34dc6eb
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>