kanas: bluetooth: guess stock's libbt_vndcfg config
The bluetooth works but sometimes it would just fail silently and appear
that it seemingly works. One manifestation of this is how the bluetooth
devices sometimes may not appear on the list even though after turning the
bluetooth on and off will eventually show them.

This patch is random guesses at how the stock configures it:
  1. The FW patch settlement time is doubled giving more time for the
     patch to apply
  2. The Bluetooth hardware seems to need SCO_USE_I2S_INTERFACE
     with it's parameters set to 1, 0, 0 and 0, in that specific order
  3. Based on TeamMargarine's sprd device tree, BTA_DISABLE_DELAY is about 5x
     larger than default.

Doing a short test, the bluetooth stack with these changes now correctly displays
the avialable devices most, if not all, of the times.

Change-Id: I04f6cf9fb8edf54f0ba3dd62be177cd4c76113c5
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>