kanas: explicitly include mkf2fsuserimg.sh to fix compilation regarding f2fs /data partitions
For some reason, the build will fail with

"OSError: [Errno 2] No such file or directory"

when trying to execute mkf2fsuserimg.sh.

To fix this, try explicitly including mkf2fsuserimg.sh. The target is a host
binary and will not appear in the image so it should be safe to do this
even though the builder does not use/need/want an f2fs /data/ partition.

I don't even know why this happens to me even though I use ext4 for the /data/ partition.

Change-Id: I33626b5a8e216f63a601a6afc77948f0e9fa1f02
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>