Revert "kanas: overlays: Reduce LMK values"
This reverts commit 254ef31cea795a7f22d467c5ba0f58f6606a7fce.

There's a kernel patch to fix up the ZRAM awareness of in-kernel LMK.
That awareness basically uses the ZRAM capacity as a metric to
dynamically adjust the adjustment score of the process to be killed.

This reduction of LMK values patch's main goal is to prevent LMK from killing
too much of background applications to the point where even going to the
recents screen have a high probablity of
killing the app you've just used.

With this revert, LMK may fire up earlier but will target higher adj scored
processes instead of the apps on the 200-300 level.

Change-Id: I171ae1077951515d2eaad679c10b9a49c19a5ccd
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>