kanas: overlays: Adjust LMK values
Previous attempt at doing this is increase the free memory by reducing the
likeliness LMK will run. But acctually, reducing it to 48MB will also reduce
it's effectiveness and its reason for existence. As a refresher, LMK was written
to prevent OOM scenarios by doing what an OOM-killer does much much before
it can happen.

And another consecuence is that it reduces the memory used for caching which is
actually "that" important when doing some real work with this device. With a
smaller cache, there may be more apps open but it will perform noticably slower.

There are more side-effects for reducing LMK values, primarily due to having smaller
cache and free memory. For now, set it to 100MB.

Change-Id: I010c15732fb235725e1c96aafd295716d76b993e
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>