kanas: Re-enable cpusets and disable hotplugs
This reverts commit 4c9b698ced52ff37e412aecc8633362a7447daea
"Enable hotplug and remove the misunderstood flag"

Based on my initial informal testing, it seems that cpu hotplugging
does no significant powersaving benefits. It increases scheduling
overhead, delay and latency whenever a core is brought up and down.

On a semi-blind testing, hotplugging does seem to save a bit of juice but
it also seemed to be just a trade off eith performance since
the core could be down while the device is experiencing load
and when it is brought up, the load may have been already done.

On idle testing (I left the phone with screen off but with a wakelock),
I measured the temperature. It showed no noticable (or quantifiable)
difference between having all 4 cores active and a single core active.
The temparature is close to the ambient temperature of my location.

With this simple highly unscientific and debatable tests,
this post from https://elementalx.org/the-truth-about-kernels-and-battery-life/ ,
and this https://source.android.com/devices/tech/debug/jank_capacity .
I had an idea and a reason to revisit this issue.

To explore cpu capacity related issues.

Though this means that, upon enabling cpusets, cpu hotplugging may be disabled.
Fortunately, it seems that someone has made a kernel patch just for that.

Revert when experiencing issues

Change-Id: Ib7fde10a4f1b3700ff0cdec2ad9b85dd6dc3ff46
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>