kanas: gps: Add back the modified gps.h header
This concerns the wrong assumption on the commit
986606a6eb517006d5ec2f9096b5e44c3f9e5673 (kanas: Add "/dev/ion (deleted)" as one of the whitelisted paths in zygote; Do not use scx35-common's custom gps.h header)

At least on the second part:
 > The second part pertains to the gps.h being used as a part of the common config
 > from scx35-common. We're using the gps hal wrapper so it shouldn't matter if
 > we are using the custom or the original header.

Apparently, the Marvell GPS lib that came in with the stock uses some
Samsung changes that *WILL* mess up the reports to the Android system since
it changes some structs.
Not using the modified gps.h will result in some funky output like
the some satellites having some -28373889 id or with 0.0db SNR. This will also
prevent showing which satellites are being used in fix and
in theory should show more satellites (though invalid) than what the GPS HAL truly
reports.

Though it shouldn't affect the performance (from what it looks like when I'm testing it)
and the GPS HAL still reports the right location.

The fix should be simple (revert), but to keep the first part about /dev/ion,
create a copy of the modified gps.h header on our device tree.

Change-Id: I72fb3eea4b04309b147cefdda0f44ca09c93c777
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>