kanas: Tweak art a bit in attempt to reduce memory usage
dex2oat default filter for system apps is now set to "balanced" instead
of "speed" (to reduce the size taken up on /data)

Target utilization of the heap is increased from 70% to 90% which
should lessen the ram wasted in allocations
but causes GC to run more frequently.

ART Heapsize is reduced to 96M and should help some large apps against
allocating some absurd (in this device context) amounts of memory just
for the sake of optimizations.

Correct me if I'm wrong; I'm still testing these.

Change-Id: Ie50dc5da7f9abe28c7c6842a5b532df52e036f32
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>