kanas: libsensors: Stub SMD feature
This commit should effectively turn SMD feature to no-op. The SMD
sensor is not really removed but enabling it from Android will
not cause any SMD triggers (let alone enabling the accelerometer).

The reason?
The smd feature that was based upon the reactive alert flag in the
kernel DOES NOT WAKE the device, meaning that the SMD fails to
serve its purpose as a wakeup sensor (and in turn just wastes
energy since the accelerometer is still powered). But, we'd still like
to have the Doze feature active.

Change-Id: Ifaa0e04ccc3f93ee3a07fdbe512db83b40b46c01
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
(cherry picked from commit 850ef9ad1ea9a420d187a5f1ba419fb8cecda51f)
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>