kanas: audio: Slightly Increase speaker volume
This follows the scx35-common commit
1401f0ab991780c294c11dcfa61ea334f91dd916 (audio: Increase speaker and headphones volume)

The key difference here is that by slightly altering the values
we should have at least a half increase of the speaker volume without
the speaker clipping the output.

Headphones aren't affected.

NOTE: The values may be devices dependent; you can build the tinymix binary
to check if the values are what you expect.

Change-Id: Icbde0a5a82d6a3d3e2813e01cddad48b7e42ee60
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>