kanas: media_profile.xml: Add 720p profile for the camcoder 0
This should now allow 720p recording but with a catch.

The stock camera driver will only allow up to 640x480 framesize and
thus be stuck with that size. The opensourced driver could use the
1280x720 framesize but you lose most of the features of the stock driver.

Both drivers will still produce a 720p video but the catch is:
with the 640x480 frame size of the stock driver, the GSP module
is forced to do an Upscale which may cause an overflow and extreme
delays in processing. This is evident when you used the zoom.
This explains the 960x720 resolution instead of 1280x720 or 1280x960.

The opensource driver doesn't have those problem can handle
the 1280x720 video res.

Next patch should help fix the wrong audio pitch/speed on screenrecords
and stuff.

Change-Id: I5e1b2a76faf673012caf9b88bc77eec1f9967727
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
(cherry picked from commit 73229133f608bacb33ce183404f40e3162a921d4)
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>