kanas: patches: Update sprd components
One patch should update the MP3 module and the other would allow us to use
the scx30g2 avcdec module which works much better than the scx15.

WARNING: because of the scx30g2 avcdec patch, the vendor tree should have been
updated with a binary patched libomx* library. The binary patched library
will invert the frequency table since the table that the kernel uses
is the reverse of the newer libomx library.

The MP3 module, as stated in the patch, should now have an expllicit support
for mp2 and mp1 files. (I only have one mp2 file disguised as an mp3 so
it's a win situation for me :P)

Change-Id: Id9074d37cfec102b7273da424a21a430b11f6876
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>