kanas: bluetooth: Apply the bluedroid header; Remove BTM_DEF_LOCAL_NAME
This was supposed to added during the 03290b068b0f9449ae9ce53de48b3281399e4149
cherry pick but it was left out.

Regarding the BTM_DEF_LOCAL_NAME, I could find the actual default name in
the stock rom so I assume they used something else not hardcoded to the
library. My guess is the AOSP default which is derived from the product name.

Change-Id: If01090c99a959b19356f349571e86f2301520835
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>