kanas: Update HIDL configuration
From the scx35-common, there are a lot of stuff we can already use with HIDL
especially with the Gralloc workarounds it should be possible to include it here
BUT as a passthrough. It will fail otherwise.

This patch would try to enable all of the services that I've tested to work on
kanas3g.

Change-Id: I28009e3b401fcd373081381f907ddb93dd1dda47
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>