kanas: fix permission problems of accel, camera, audio, and gsp
The libsensors HAL cannot open the event fd for the accelerometer
so screen rotation will not work because the process using the HAL
has a different group. The same goes for the audio, though it's not
a show stopper but it prevents the DA EQ from ever activating because
it cannot write to /data/local/media.

On the graphics side, the sprd_gsp module uses some absurd permission
by comparison to the Android N tree. I do not know how the scx35-common
guys got there but it seems that we need to override it again.

Change-Id: I2a49094ea16eeff960e4db96d85c2d36abf7021a
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>