kanas: audio: increase volume steps to 25
This is of personal desire to have a quiter volume since there are
some music I often listened to that are still relatively loud even when
the volume is set to 1. "Relative" means being relative to the surroundings
(i.e. when the surroundings are too quiet, even a classical music piece may
seem too loud)

PS. I usually use 60 instead of 25 but it brings an inconvinience of having
to smash the volume rocker 2-3x more times just to addjust the volume.

Change-Id: I7ff97cc0b06949a99387f14cc908ba7701e1c321
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>