kanas: Autostart refnotify
This is a seemingly innocent software that when left unactivated
will cause a memory leak on the WCDMA processor (or co-proc or just simply
the modem, I don't care).

The stock configs also do start refnotify independently of the
modemd. Though, on other stock roms of other SPRD+SAMSUNG phones
it's being started manually (not by init)
 either by modemd or some on property set event.

Change-Id: I99b98b14f8ca84ceb9a5adbd40bf296812cd02dd
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>