kanas: ril: add additional init scripts to support the scx15 of modemd
I am still testing the sc8830 modemd and so far the only thing that it
cannot do is to properly reset the phone...

First, it does not issue the kill_*phone service correctly.
In scx15 it's kill_*_phone where * is one of the three {w, t, l}
In sc8830 it's just kill_phone

Second, For some reason the modem_control is the one that will do the modem
reset, including the reflash and stuff but it somehow cannot do it leaving
the cell radio in a bricked state.

This commit will only solve the First issue. To solve the second, we either
have to use the scx15 version or fix the sc8830 version of modemd.
In reality, just using the scx15 version would be better since the vendor
blobs are already the scx15.

This commit actually just helps either versions to run on the same config
without needing to have <<Init Wizardry>> skill. Just slap the appropiate
binaries then test.

Change-Id: I65884fae0436d40f98353671d169c02a3788cbe0
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>